setup one entrypoint for the agentbakersvc.
